chore: rename frontend.server.* to frontend.*
