+++ src/app/modules/frontend/google-analytics.factory.ts
@@ -7,7 +7,7 @@ import FeatureManager, {
   RegisteredFeature,
 } from '../../../config/feature-manager'
 
-import * as FrontendServerController from './frontend.server.controller'
+import * as FrontendServerController from './frontend.controller'
 
 interface IGoogleAnalyticsFactory {
   addGoogleAnalyticsData: RequestHandler<
