+++ src/loaders/express/index.ts
@@ -15,7 +15,7 @@ import { BounceRouter } from '../../app/modules/bounce/bounce.routes'
 import { ExamplesRouter } from '../../app/modules/examples/examples.routes'
 import { AdminFormsRouter } from '../../app/modules/form/admin-form/admin-form.routes'
 import { PublicFormRouter } from '../../app/modules/form/public-form/public-form.routes'
-import { FrontendRouter } from '../../app/modules/frontend/frontend.server.routes'
+import { FrontendRouter } from '../../app/modules/frontend/frontend.routes'
 import { HomeRouter } from '../../app/modules/home/home.routes'
 import { MYINFO_ROUTER_PREFIX } from '../../app/modules/myinfo/myinfo.constants'
 import { MyInfoRouter } from '../../app/modules/myinfo/myinfo.routes'
